controlcenter: move to dev/controlcenter/
