+++ .github/workflows/go.yml
@@ -16,9 +16,9 @@ jobs:
           cache: true
       - uses: oven-sh/setup-bun@v2
       - run: bun install
-        working-directory: controlcenter/ui
+        working-directory: dev/controlcenter/ui
       - run: bun run build
-        working-directory: controlcenter/ui
+        working-directory: dev/controlcenter/ui
       - run: go build ./...
   go-lint:
     runs-on: ubuntu-latest
@@ -31,9 +31,9 @@ jobs:
           cache: true
       - uses: oven-sh/setup-bun@v2
       - run: bun install
-        working-directory: controlcenter/ui
+        working-directory: dev/controlcenter/ui
       - run: bun run build
-        working-directory: controlcenter/ui
+        working-directory: dev/controlcenter/ui
       - uses: golangci/golangci-lint-action@v8
   go-test:
     runs-on: ubuntu-latest
@@ -46,9 +46,9 @@ jobs:
           cache: true
       - uses: oven-sh/setup-bun@v2
       - run: bun install
-        working-directory: controlcenter/ui
+        working-directory: dev/controlcenter/ui
       - run: bun run build
-        working-directory: controlcenter/ui
+        working-directory: dev/controlcenter/ui
       - run: go test ./...
   evals-short:
     runs-on: ubuntu-latest
