+++ dev/controlcenter/internal/logs/parser.go
