+++ dev/controlcenter/ui/public/favicon.ico
