+++ dev/controlcenter/ui/src/components/LogHistogram.tsx
