+++ dev/controlcenter/ui/src/components/LogViewer.tsx
