+++ dev/controlcenter/ui/src/components/Sidebar.tsx
