+++ dev/controlcenter/ui/src/pages/Logs.tsx
