+++ dev/controlcenter/ui/src/vite-env.d.ts
