+++ dev/controlcenter/ui/tsconfig.json
