+++ dev/controlcenter/ui/vite.config.ts
